ajustes movimentacao

+++ src/main/java/br/senac/cantina/produto/dto/CreateProdutoDto.java
@@ -5,8 +5,8 @@ import java.time.LocalDateTime;
 public record CreateProdutoDto (
     String nome,
     String descricao,
-    double preco,
-    float quantidade,
+    Double preco,
+    Double quantidade,
     LocalDateTime dataHora){
 
 }
